ensure float

+++ transpnerf/depth_normal_dataset.py
@@ -80,14 +80,14 @@ class DepthNormalDataset(InputDataset):
                                         [-1.00000, 0.00000, 1.00000],
                                         [0.00000, 0.00000, 0.00000]])
         delzdelxkernel = delzdelxkernel.view(1, 1, 3, 3).repeat(1, nb_channels, 1, 1)
-        delzdelx = F.conv2d(depths_reshape, delzdelxkernel)
+        delzdelx = F.conv2d(depths_reshape, delzdelxkernel, dtype=torch.float64)
 
         delzdelykernel = torch.tensor([[0.00000, -1.00000, 0.00000],
                                         [0.00000, 0.00000, 0.00000],
                                         [0.0000, 1.00000, 0.00000]])
         delzdelykernel = delzdelykernel.view(1, 1, 3, 3).repeat(1, nb_channels, 1, 1)
 
-        delzdely = F.conv2d(depths_reshape, delzdelykernel)
+        delzdely = F.conv2d(depths_reshape, delzdelykernel, dtype=torch.float64)
 
         delzdelz = torch.ones(delzdely.shape, dtype=torch.float64)
